+++ web/src/frontend/scripts/mock_data/soknadsosialhjelp_nb_NO.properties
@@ -182,6 +182,8 @@ personalia.bosted.oppsummering=Når du har fylt ut blir søknaden sendt til
 personalia.bydel.label=I hvilken bydel?
 personalia.kommune.label=I hvilken kommune bor du og-/eller oppholder deg i?
 skjema.tittel=Søknad for økonomisk sosialhjelp
+skjema.feilmeldinger.feil=Feil
+skjema.feilmeldinger.404=404 Ukjent side
 utgifter.barn.false=Nei
 utgifter.barn.sporsmal=Har du utgifter til barn?
 utgifter.barn.true.utgifter.annet.true.beskrivelse.label=Beskriv
