+++ web/src/frontend/src/nav-soknad/faktum/StegFaktum.tsx
@@ -5,6 +5,7 @@ import { getIntlTextOrKey } from "../utils";
 import Feiloppsummering from "../components/validering/Feiloppsummering";
 import { SoknadAppState } from "../redux/faktaReducer";
 import { Valideringsfeil } from "../validering/types";
+import StegMedNavigasjon from "../../digisos/skjema/stegMedNavigasjon";
 const DocumentTitle = require("react-document-title");
 
 interface OwnProps extends React.Props<any> {
@@ -39,20 +40,22 @@ class StegFaktum extends React.Component<
 			+ getIntlTextOrKey(intl, tittelId);
 		return (
 			<DocumentTitle title={title}>
-				<div className="skjema-steg skjema-content">
-					<div className="skjema-steg__feiloppsummering">
-						<Feiloppsummering
-							skjemanavn="digisos"
-							valideringsfeil={valideringsfeil}
-							stegValidertCounter={stegValidertCounter}
-							visFeilliste={visFeilmeldinger}
-						/>
+				<StegMedNavigasjon>
+					<div className="skjema-steg skjema-content">
+						<div className="skjema-steg__feiloppsummering">
+							<Feiloppsummering
+								skjemanavn="digisos"
+								valideringsfeil={valideringsfeil}
+								stegValidertCounter={stegValidertCounter}
+								visFeilliste={visFeilmeldinger}
+							/>
+						</div>
+						<h2 className="skjema-steg__tittel" ref={c => (this.tittel = c)}>
+							{getIntlTextOrKey(intl, tittelId)}
+						</h2>
+						{children}
 					</div>
-					<h2 className="skjema-steg__tittel" ref={c => (this.tittel = c)}>
-						{getIntlTextOrKey(intl, tittelId)}
-					</h2>
-					{children}
-				</div>
+				</StegMedNavigasjon>
 			</DocumentTitle>
 		);
 	}
